requirements.txt
Pillow==11.0.0
pygame==2.6.1
win10toast==0.9
